update vsync with daptive option by default ensure DPI scaling creates correct window size remove old resource path code on macOS fix window event messages and camera callbacks

+++ crogine/include/crogine/core/Message.hpp
@@ -91,7 +91,7 @@ namespace cro
         struct WindowEvent final
         {
             std::uint32_t windowID = 0;
-            std::uint8_t event = 0;
+            std::uint32_t event = 0;
             std::int32_t data0 = 0;
             std::int32_t data1 = 0;
         };
